+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ### Changes
 - ItemViewModel.bindTo(itemFragment) supports all item fragments now, not just ListCellFragment
 - lambda's that return unit are no longer nullable. use the default-lambda instead
-
+- ChildInterceptor is now an Interface. 
 ### Additions
 - `cellFragment` support for DataGrid
 - ObservableValue<String>.isBlank() and ObservableValue<String>.isNotBlank() which returns BooleanBinding. Useful for binding to TextField enabled/visible state